# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VJU Club Management API
  version: 1.0.0
  description: 'API giai đoạn 1 cho quản lý tài khoản, CLB, ban, thành viên và permission.
    Xem x-implementation-status từng operation: IMPLEMENTED là đã xác nhận qua integration
    test. UUID, camelCase, thời gian ISO-8601. List dùng offset/limit; không có page/size.'
servers:
- url: /api/v1
  description: Backend cùng origin với Swagger UI
- url: http://localhost:8080/api/v1
  description: Backend local mặc định
security:
- sessionAuth: []
paths:
...
  /clubs/{clubId}/documents:
    get:
      operationId: document.list
      summary: List club documents
      tags:
      - Document
      x-permission: document.view
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: appDetailKey
        in: query
        required: false
        schema:
          type: string
      - name: deleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'Mới tạo trước. deleted=true (thùng rác) cần thêm document.restore.'
    post:
      operationId: document.upload
      summary: Upload a document
      tags:
      - Document
      x-permission: document.upload
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      description: 'File lưu trên R2 dưới năm/tháng/ngày/uuid.ext. Tên, đuôi và nội dung file được kiểm tra; trùng tên trả 409 DOCUMENT_NAME_ALREADY_EXISTS.'
  /clubs/{clubId}/documents/count:
    get:
      operationId: document.count
      summary: Count club documents
      tags:
      - Document
      x-permission: document.view
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: appDetailKey
        in: query
        required: false
        schema:
          type: string
      - name: deleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'getDocCount.'
  /clubs/{clubId}/documents/names:
    get:
      operationId: document.names
      summary: List document names
      tags:
      - Document
      x-permission: document.view
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: deleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentNames'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'getDocumentNames, sắp xếp A–Z.'
  /documents/{documentId}:
    get:
      operationId: document.get
      summary: Get a document
      tags:
      - Document
      x-permission: document.view
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'Tài liệu đã xoá chỉ hiện với người có document.restore.'
    patch:
      operationId: document.update
      summary: Rename a document
      tags:
      - Document
      x-permission: document.update
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentPatchRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      description: 'Chủ sở hữu hoặc document.update. Tên mới phải giữ đuôi file; appDetailKey rỗng để xoá.'
    delete:
      operationId: document.delete
      summary: Soft-delete a document
      tags:
      - Document
      x-permission: document.delete
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'Xoá mềm: đánh dấu deleted, file vẫn còn trên R2. Chủ sở hữu hoặc document.delete.'
  /documents/{documentId}/restore:
    post:
      operationId: document.restore
      summary: Restore a deleted document
      tags:
      - Document
      x-permission: document.restore
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      description: 'Khôi phục tài liệu đã xoá mềm; tên đã bị dùng trả 409.'
  /documents/{documentId}/versions:
    get:
      operationId: document.versions
      summary: List document versions
      tags:
      - Document
      x-permission: document.view
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'Mới nhất trước.'
    post:
      operationId: document.upload_version
      summary: Upload a new version
      tags:
      - Document
      x-permission: document.upload
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentVersionUploadRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      description: 'Cùng loại file, khác nội dung phiên bản hiện tại; expectedVersion cũ trả 409 DOCUMENT_VERSION_CONFLICT.'
  /documents/{documentId}/download:
    get:
      operationId: document.download
      summary: Download a document
      tags:
      - Document
      x-permission: document.view
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: version
        in: query
        required: false
        description: Bỏ trống = phiên bản hiện tại
        schema:
          type: integer
      responses:
        '200':
          description: File content (Content-Disposition attachment, ETag = SHA-256)
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'Stream file qua backend.'
  /documents/{documentId}/download-url:
    get:
      operationId: document.download_url
      summary: Get a presigned download link
      tags:
      - Document
      x-permission: document.view
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: version
        in: query
        required: false
        description: Bỏ trống = phiên bản hiện tại
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDownloadUrl'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'Link R2 ký sẵn, hết hạn sau 5 phút; storage local trả đường dẫn /download.'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: CLUB_SESSION
      description: 'Cookie phiên do POST /auth/login đặt. Request thay đổi dữ liệu cần thêm header X-XSRF-TOKEN.'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 'Chưa đăng nhập hoặc phiên đã hết hạn/bị huỷ (code UNAUTHORIZED); đăng nhập lại.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Permission denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Trùng dữ liệu hoặc trạng thái hiện tại không cho phép thao tác.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Không tìm thấy resource hoặc grant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
          description: URI request, có thể vắng với lỗi do security filter trả về.
        code:
          type: string
      required:
      - type
      - title
      - status
      - detail
      - code
      examples:
      - type: about:blank
        title: PERMISSION_DENIED
        status: 403
        detail: Permission denied
        instance: /api/v1/clubs/22222222-2222-4222-8222-222222222222
        code: PERMISSION_DENIED
...
    Document:
      type: object
      required: [id, clubId, ownerId, name, path, version, contentType, sizeBytes, checksumSha256, deleted, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        clubId: {type: string, format: uuid}
        ownerId: {type: string, format: uuid}
        name: {type: string}
        path: {type: string, description: 'Object key trên R2: yyyy/MM/dd/uuid.ext'}
        version: {type: integer}
        contentType: {type: string}
        sizeBytes: {type: integer, format: int64}
        checksumSha256: {type: string}
        appDetailKey: {type: [string, 'null']}
        deleted: {type: boolean}
        deletedAt: {type: [string, 'null'], format: date-time}
        deletedBy: {type: [string, 'null'], format: uuid}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    DocumentListResponse:
      type: object
      required: [items, total, offset, limit]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        total: {type: integer, format: int64}
        offset: {type: integer}
        limit: {type: integer}
    DocumentVersion:
      type: object
      properties:
        id: {type: string, format: uuid}
        documentId: {type: string, format: uuid}
        version: {type: integer}
        originalName: {type: string}
        contentType: {type: string}
        sizeBytes: {type: integer, format: int64}
        checksumSha256: {type: string}
        uploadedBy: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    DocumentCount:
      type: object
      properties:
        clubId: {type: string, format: uuid}
        deleted: {type: boolean}
        count: {type: integer, format: int64}
    DocumentNames:
      type: object
      properties:
        clubId: {type: string, format: uuid}
        deleted: {type: boolean}
        names:
          type: array
          items: {type: string}
    DocumentDownloadUrl:
      type: object
      properties:
        url: {type: string}
        version: {type: integer}
        expiresAt: {type: [string, 'null'], format: date-time}
    DocumentPatchRequest:
      type: object
      properties:
        name: {type: string, maxLength: 255}
        appDetailKey: {type: string, maxLength: 100}
    DocumentUploadRequest:
      type: object
      required: [file]
      properties:
        file: {type: string, format: binary}
        name: {type: string}
        appDetailKey: {type: string}
    DocumentVersionUploadRequest:
      type: object
      required: [file]
      properties:
        file: {type: string, format: binary}
        expectedVersion: {type: integer}
...